+++ app/controllers/facilities_management/rm6378/admin/frameworks_controller.rb
@@ -1,0 +1,9 @@
+module FacilitiesManagement
+  module RM6378
+    module Admin
+      class FrameworksController < FacilitiesManagement::Admin::FrameworkController
+        include ::Admin::FrameworksController
+      end
+    end
+  end
+end