+++ app/controllers/facilities_management/rm6378/admin/reports_controller.rb
@@ -1,0 +1,9 @@
+module FacilitiesManagement
+  module RM6378
+    module Admin
+      class ReportsController < FacilitiesManagement::Admin::FrameworkController
+        include ::Admin::ReportsController
+      end
+    end
+  end
+end