+++ app/models/user.rb
@@ -59,6 +59,7 @@ class User < ApplicationRecord
   has_many :searches, inverse_of: :user, dependent: :destroy
   has_many :reports, inverse_of: :user, dependent: :destroy
   has_many :procurements, inverse_of: :user, dependent: :destroy
+  has_many :admin_uploads, inverse_of: :user, class_name: 'Admin::Upload', dependent: :destroy
 
   # Include default devise modules. Others available are:
   # :confirmable, :lockable, :trackable and :omniauthable