+++ features/step_definitions/facilities_management/admin_steps.rb
@@ -13,6 +13,15 @@ Given('I sign in as an admin and navigate to the {string} dashboard') do |framew
   expect(page.find('h1')).to have_text("#{framework} administration dashboard")
 end
 
+Given('I sign in as an admin for the {string} framework in {string}') do |framework, _service|
+  @framework = framework
+  visit "/facilities-management/#{framework}/admin/sign-in"
+  update_banner_cookie(true) if @javascript
+  create_admin_user_with_details
+  step 'I sign in'
+  step "I am on the 'Admin dashboard' page"
+end
+
 Given('I go to the facilities management {string} admin start page') do |framework|
   visit "/facilities-management/#{framework}/admin/sign-in"
   update_banner_cookie(true) if @javascript
@@ -48,8 +57,63 @@ Then('there should be {int} management reports') do |number_of_management_report
   expect(admin_page.management_reports.length).to eq number_of_management_reports
 end
 
+Then('I should see the following details in the {string} summary:') do |summary, supplier_data|
+  summary_rows = admin_page.send(summary)
+
+  expect(summary_rows.length).to eq(supplier_data.raw.length)
+
+  summary_rows.zip(supplier_data.raw).each do |section, (expected_key, expected_value)|
+    expect(section.key).to have_text(expected_key)
+    expect(section.value).to have_text(expected_value)
+  end
+end
+
+Then('I should see the following details in the summary for the lot {string}:') do |lot_name, supplier_data|
+  summary_rows = admin_page.supplier_lots.find { |element| element.lot_name.text == lot_name }.lot_info
+
+  expect(summary_rows.length).to eq(supplier_data.raw.length)
+
+  summary_rows.zip(supplier_data.raw).each do |section, (expected_key, expected_value)|
+    expect(section.key).to have_text(expected_key)
+    expect(section.value).to have_text(expected_value)
+  end
+end
+
+Then('I click on {string} for the lot {string}') do |view_link_text, lot_name|
+  admin_page.supplier_lots.find { |element| element.lot_name.text == lot_name }.lot_info.find { |section| section.value.text.starts_with?(view_link_text) }.find('a').click
+end
+
+Then('the supplier should be assigned to the {string} as follows:') do |_section, section_items|
+  admin_check_section_items(
+    admin_page.supplier_section_summaries.first.section_items.first,
+    section_items
+  )
+end
+
+Then('the supplier should be assigned to the {string} in {string} as follows:') do |_section, category, section_items|
+  admin_check_section_items(
+    admin_page.supplier_section_summaries.first.section_items.find { |section| section.heading.text == category },
+    section_items
+  )
+end
+
+Then('the supplier should not be assigned any {string} with the following message:') do |_section, empty_message|
+  expect(admin_page.supplier_section_summaries.first.empty_message).to have_text(empty_message.raw.flatten.first)
+end
+
 def add_management_report_dates(date_type, day, month, year)
   admin_page.management_report.send("#{date_type} day").set(day)
   admin_page.management_report.send("#{date_type} month").set(month)
   admin_page.management_report.send("#{date_type} year").set(year)
+end
+
+def admin_check_section_items(summary, section_items)
+  items = summary.items
+  expected_items = section_items.raw.flatten.compact.compact_blank
+
+  expect(items.length).to eq(expected_items.length)
+
+  items.zip(expected_items).each do |item, expected_value|
+    expect(item).to have_text(expected_value)
+  end
 end