+++ features/support/pages/rm6378/admin.rb
@@ -1,0 +1,6 @@
+require_relative '../admin'
+
+module Pages::RM6378
+  class Admin < Pages::Admin
+  end
+end