+++ spec/factories/facilities_management/rm6378/admin/supplier.rb
@@ -1,0 +1,3 @@
+FactoryBot.define do
+  factory :facilities_management_rm6378_admin_supplier, class: 'FacilitiesManagement::RM6378::Admin::Supplier', parent: :supplier
+end