+++ spec/models/supplier/framework/lot_spec.rb
@@ -1,9 +1,9 @@
 require 'rails_helper'
 
 RSpec.describe Supplier::Framework::Lot do
-  describe 'associations' do
-    let(:supplier_framework_lot) { create(:supplier_framework_lot) }
+  let(:supplier_framework_lot) { create(:supplier_framework_lot) }
 
+  describe 'associations' do
     it { is_expected.to belong_to(:supplier_framework) }
     it { is_expected.to belong_to(:lot) }
 
@@ -31,4 +31,50 @@ RSpec.describe Supplier::Framework::Lot do
       expect { create(:supplier_framework_lot, supplier_framework:, lot:) }.to raise_error(ActiveRecord::RecordNotUnique)
     end
   end
+
+  # rubocop:disable RSpec/NestedGroups
+  describe 'validations' do
+    before { supplier_framework_lot.assign_attributes(attributes) }
+
+    context 'when validating lot_status' do
+      let(:attributes) { { enabled: } }
+
+      let(:enabled) { true }
+
+      context 'when considering the enabled status' do
+        context 'and it is nil' do
+          let(:enabled) { nil }
+
+          it 'is invalid and has the correct error message' do
+            expect(supplier_framework_lot.valid?(:lot_status)).to be(false)
+            expect(supplier_framework_lot.errors[:enabled].first).to eq('Select if the supplier is on this lot')
+          end
+        end
+
+        context 'and it is blank' do
+          let(:enabled) { '' }
+
+          it 'is invalid and has the correct error message' do
+            expect(supplier_framework_lot.valid?(:lot_status)).to be(false)
+            expect(supplier_framework_lot.errors[:enabled].first).to eq('Select if the supplier is on this lot')
+          end
+        end
+
+        context 'and it is true' do
+          it 'is valid' do
+            expect(supplier_framework_lot).to be_valid(:lot_status)
+          end
+        end
+
+        context 'and it is false' do
+          let(:enabled) { false }
+
+          it 'is valid' do
+            expect(supplier_framework_lot).to be_valid(:lot_status)
+          end
+        end
+      end
+    end
+  end
+  # rubocop:enable RSpec/NestedGroups
 end