+++ CHANGELOG.md
@@ -7,11 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add the inital admin pages for Facilities Management RM6378 ([PR 5318](https://github.com/Crown-Commercial-Service/crown-marketplace/pull/5318))
+- Add the managing of supplier and lot data for Facilities Management RM6378 ([PR 5318](https://github.com/Crown-Commercial-Service/crown-marketplace/pull/5318))
+
 ### Changed
 
 - Update CCS Frontend to v3.2.0 ([PR 5276](https://github.com/Crown-Commercial-Service/crown-marketplace/pull/5276))
 - Remove the "use gca" flags ([PR 5276](https://github.com/Crown-Commercial-Service/crown-marketplace/pull/5276))
-- Update how we do releases to preview and production to target a specific tag  ([PR 5308](https://github.com/Crown-Commercial-Service/crown-marketplace/pull/5308))
+- Update how we do releases to preview and production to target a specific tag ([PR 5308](https://github.com/Crown-Commercial-Service/crown-marketplace/pull/5308))
+- Allow framework status to be updated at the framework level rather than the service level ([PR 5318](https://github.com/Crown-Commercial-Service/crown-marketplace/pull/5318))
 
 ## [7.7.0] - 2026-04-30